fix: 좋아요, 조회수, 공유 컴포넌트 분리
